adding flags diouxX with specifier hh and h

+++ src/ft_outputshort.c
@@ -6,7 +6,7 @@ void    ft_outputshort(va_list args, int space, char c)
     long long i;
 
     if (c == 'l')
-        i = va_arg(args, long);
+        i = va_arg(args, int);
     else    
         i = va_arg(args, long long);
     ft_putnbrlong(i);
@@ -18,7 +18,7 @@ void    ft_outputshorthex(va_list args, int space, char hex, char c)
     (void)hex;
     long long i;
     if (c == 'l')
-        i = va_arg(args, long);
+        i = va_arg(args, int);
     else    
         i = va_arg(args, long long);
     ft_putstr(ft_longtoa_base(i, 16, hex));
